pyproject.toml
[tool.poetry]
name = "taipei-day-trip"
version = "0.1.0"
description = ""
authors = ["youyuhsuan <stellayou0118@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
mysql-connector-python = "^8.4.0"
uvicorn = "^0.30.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"


[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
